requirements.txt
python-dotenv==1.0.1
langchain==0.2.6
langchain-community==0.2.6
langchain-openai==0.1.13
langchain-google-genai==1.0.6
langgraph==0.1.5
fastapi==0.111.0
uvicorn==0.30.1
google-api-python-client==2.135.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
beautifulsoup4==4.12.3
twilio==9.2.0
tavily-python==0.3.3
pydantic==2.8.2
requests==2.32.3
duckduckgo-search==6.1.7